aws-iot-device: Update package
Update commit SHA, which contains changes relate to how the library gets
installed, such as installing the library files into the /usr/lib/aws
directory, and installing the 3rd-party libraries mbedtls and tinycbor as
part of the install process.

This commit also sets the runtime paths of the installed libraries.